+++ ShoppingListApi/Model/DTOs/Get/UserRoleGetDto.cs
@@ -1,4 +1,6 @@
-﻿namespace ShoppingListApi.Model.DTOs.Get;
+﻿using ShoppingListApi.Model.Entity;
+
+namespace ShoppingListApi.Model.DTOs.Get;
 
 public class UserRoleGetDto
 {
@@ -11,5 +13,15 @@ public class UserRoleGetDto
         UserRoleId = userRoleId;
         UserRoleTitle = userRoleTitle;
         EnumIndex = enumIndex;
+    }
+    
+    public static UserRoleGetDto FromUserRole(UserRole userRole)
+    {
+        return new UserRoleGetDto(userRole.UserRoleId, userRole.UserRoleTitle, userRole.EnumIndex);
+    }
+    
+    public static List<UserRoleGetDto> FromUserRoleList(List<UserRole> userRoles)
+    {
+        return userRoles.Select(FromUserRole).ToList();
     }
 }